feat(sb): added more endpoints
