+++ apps/sandbox/package.json
@@ -11,7 +11,10 @@
     "hono": "^4.6.4",
     "hono-openapi": "^0.1.5",
     "zod": "^3.23.8",
-    "zod-openapi": "^4.0.0"
+    "zod-openapi": "^4.0.0",
+    "valibot": "^1.0.0-beta.9",
+    "arktype": "2.0.0-rc.25",
+    "@sinclair/typebox": "^0.34.9"
   },
   "devDependencies": {
     "@types/node": "^20.11.17",
